Made changes for code coverage
