+++ MetaMorpheus/EngineLayer/PeptideSpectralMatch.cs
@@ -56,20 +56,4 @@ namespace EngineLayer
         #endregion
 
     }
-
-    public class OligoSpectralMatch : SpectralMatch
-    {
-        public OligoSpectralMatch(IBioPolymerWithSetMods peptide, int notch, double score, int scanIndex,
-            Ms2ScanWithSpecificMass scan, CommonParameters commonParameters,
-            List<MatchedFragmentIon> matchedFragmentIons, double xcorr = 0) : base(peptide, notch, score, scanIndex,
-            scan, commonParameters, matchedFragmentIons, xcorr)
-        {
-
-        }
-
-        protected OligoSpectralMatch(SpectralMatch psm, List<(int Notch, IBioPolymerWithSetMods Peptide)> bestMatchingPeptides)
-            : base(psm, bestMatchingPeptides)
-        {
-        }
-    }
 }
